--- container_fundamentals/07_dockerfile/Dockerfile
@@ -1,2 +1,0 @@
-FROM ubuntu:20.10
-ENTRYPOINT [ "/bin/sh", "-c", "while true; do echo $(date); sleep 1; done;" ]
