--- container_fundamentals/08_dockerfile_webserver/Dockerfile
@@ -1,4 +1,0 @@
-FROM ubuntu:20.10
-RUN apt-get update && apt-get install apache2 -y && apt-get clean
-COPY index.html /var/www/html/index.html
-CMD [ "apache2ctl", "-DFOREGROUND" ]
