--- container_fundamentals/09_build-ignore/Dockerfile
@@ -1,3 +1,0 @@
-FROM alpine:3.12.0
-ADD . /app
-CMD [ "ls", "-alh", "/app/" ]
