+++ container_fundamentals/09_dockerfile/.dockerignore
@@ -1,0 +1,3 @@
+README.md
+.gitignore
+main.go
