--- container_fundamentals/10_entrypoint-vs-cmd/Dockerfile
@@ -1,2 +1,0 @@
-FROM ubuntu:20.10
-CMD [ "echo", "hello docker" ]
