--- container_fundamentals/11_shell-vs-exec-form-variable-substitution/Dockerfile
@@ -1,3 +1,0 @@
-FROM ubuntu:20.10
-ENV FOO bar
-ENTRYPOINT [ "/bin/echo", "$FOO" ]
