--- container_fundamentals/12_shell-vs-exec-form-PID1/Dockerfile
@@ -1,2 +1,0 @@
-FROM ubuntu:20.10
-ENTRYPOINT [ "ps", "aux" ]
