--- container_fundamentals/13_caching/.gitignore
@@ -1,1 +1,0 @@
-node_modules
