--- container_fundamentals/13_caching/Dockerfile
@@ -1,5 +1,0 @@
-FROM node:12
-WORKDIR /app
-COPY . .
-RUN npm install
-CMD [ "npm", "start" ]
