--- container_fundamentals/14_multistaged-builds/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (	
-	"fmt"
-)
-
-func main() {
-	fmt.Println("Hello Docker from Go")
-}
