--- container_fundamentals/18_docker-compose/prometheus.yaml
@@ -1,9 +1,0 @@
-global:
-  scrape_interval: 15s
-  evaluation_interval: 15s
-scrape_configs:
-  - job_name: "cadvisor"
-    static_configs:
-      - targets: ["cadvisor:8080"]
-        labels:
-          alias: "cadvisor"
