--- container_fundamentals/21_iptables/iptables_nginx
@@ -1,32 +1,0 @@
-Chain INPUT (policy ACCEPT)
-target     prot opt source               destination         
-
-Chain FORWARD (policy DROP)
-target     prot opt source               destination         
-DOCKER-USER  all  --  anywhere             anywhere            
-DOCKER-ISOLATION-STAGE-1  all  --  anywhere             anywhere            
-ACCEPT     all  --  anywhere             anywhere             ctstate RELATED,ESTABLISHED
-DOCKER     all  --  anywhere             anywhere            
-ACCEPT     all  --  anywhere             anywhere            
-ACCEPT     all  --  anywhere             anywhere            
-
-Chain OUTPUT (policy ACCEPT)
-target     prot opt source               destination         
-
-Chain DOCKER (1 references)
-target     prot opt source               destination         
-ACCEPT     tcp  --  anywhere             172.17.0.2           tcp dpt:http
-
-Chain DOCKER-ISOLATION-STAGE-1 (1 references)
-target     prot opt source               destination         
-DOCKER-ISOLATION-STAGE-2  all  --  anywhere             anywhere            
-RETURN     all  --  anywhere             anywhere            
-
-Chain DOCKER-ISOLATION-STAGE-2 (1 references)
-target     prot opt source               destination         
-DROP       all  --  anywhere             anywhere            
-RETURN     all  --  anywhere             anywhere            
-
-Chain DOCKER-USER (1 references)
-target     prot opt source               destination         
-RETURN     all  --  anywhere             anywhere
