--- container_fundamentals/21_iptables/stuff.md
@@ -1,6 +1,0 @@
-
-iptables -L
-
-ip link
-
-docker run -it -d -p 80:80 nginx:1.19.2
