+++ container_fundamentals/810_entrypoint-vs-cmd/Dockerfile
@@ -1,0 +1,2 @@
+FROM ubuntu:20.10
+CMD [ "echo", "hello docker" ]
