+++ container_fundamentals/811_shell-vs-exec-form-variable-substitution/Dockerfile
@@ -1,0 +1,3 @@
+FROM ubuntu:20.10
+ENV FOO bar
+ENTRYPOINT [ "/bin/echo", "$FOO" ]
