+++ container_fundamentals/812_shell-vs-exec-form-PID1/Dockerfile
@@ -1,0 +1,2 @@
+FROM ubuntu:20.10
+ENTRYPOINT [ "ps", "aux" ]
