+++ container_fundamentals/813_caching/Dockerfile
@@ -1,0 +1,5 @@
+FROM node:12
+WORKDIR /app
+COPY . .
+RUN npm install
+CMD [ "npm", "start" ]
