+++ container_fundamentals/814_multistaged-builds/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (	
+	"fmt"
+)
+
+func main() {
+	fmt.Println("Hello Docker from Go")
+}
