+++ container_fundamentals/821_iptables/stuff.md
@@ -1,0 +1,6 @@
+
+iptables -L
+
+ip link
+
+docker run -it -d -p 80:80 nginx:1.19.2
